+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Tooltip now recreates after scroll or screen resize (can be disabled in settings)
 - Improved "Open link" button detection (less false detections)
 - Improved currency & measure units detection
+- Fix for currency rates not being updated correctly
 - Other small fixes and improvements
 
 #3.3.1